requirements.txt
# Core backend
fastapi
uvicorn

# Data handling and typing
pydantic
python-dotenv

# OpenAI LLM API
openai

# Embedding model
sentence-transformers

# Vector search
faiss-cpu
numpy

# Frontend
streamlit
requests
